UI and Secure Connections
The dropdown module was previously depending on weird connected lists by
index and it was very painful to use. Switched to requiring a smart
object to say which objects are allowed, which is much more maintainable
and readable.

Also implemented the basics of the whisper packet, but that will require
making encrypted connections between the peers dynamically, as there is
the expectation of privacy in whispers. Since it's just one message that
needs to be sent, I may not need a full-on algorithm, haven't decided
yet.