Make Whispers Work
Whispers work now! Kind of odd to test on one machine, because the popup
doesn't always go to the top if the appropriate program doesn't have
focus, but it does work. Next will be setting true nodes in the
connection state.

As of right now, you can only set peers to either the largest node
(center node / broadcast node) or the smallest node (end node / listener
node). This is effectively client-server, the next step is to implement
a treenode (Or something of a better name) that would have nodes 'below'
it, like the broadcast node, but also 1 main node above it (like the
listener node)