Add Midnode Module
Solely has assertions for the metadata, because it helped me think over
how the logic is going to go. Not sure if passing information via
metadata is the way to go, but it is the best way I can think of without
referencing the instance of the module.

Perhaps using an Object... argument list will be less confusing, and
wouldn't break any current code. Would put that in the constructor of
each netmodule/logicmodule